Continuing to align with review comments to week of 7th May 2018 (#5)
* SB> I see the OAM indication material did not get incorporated
SB> from draft-bryant-detnet-mpls-dp-00. I think it needs to
SB> be here even if we have a separate OAM draft.

* SB> Similarly I think the flow agregation text from
SB> draft-bryant-detnet-mpls-dp-00 should be included.

* Corrected a couple of trivial typos.

Added text around around the use of platform lable space for the S-label

Clarified the text around the uniquness of the S-label

Added some notes to the editors about the use of the term local-id

* Replaced PREF with PRF, PEF and PROF in the definitions and abbreviations sections.

I missed a PREF and will remove in another pass.

* Removed the definition of Local-ID

Added PROEF definition

Made changes to the packet replication and elimination section and to the
packet re-ordering section to align with the splitting of the PREF functions.

* Added review comments received concerning the Agregation text that
was added, and restructured the agregation text into a single
section with sub-sections.

* Removed the references to GAL as agreed in the call of 9/May/2018

* Minor typo/English fixes that do not change the meaning of the text.

* Added MPLS over IP text provided by Andy Malis and taken from
draft-malis-detnet-ip-dp-00

* Added text in the MPLS encapsulation section saying that the
payload type is identified via the S-label.

Included a note that this is done for IPv4 and IPv6 through the
S-label and explaining why.

Added a note that the service text needs to talk about this
when it is re-written.

* Aligned sections 5.2.1 and 5.2.2 (the L2 and L3 protocol
stack outlines) to the direction of the rest of the document.

This needs more work, particularly a better description.